Implement Customer and CustomerPack classes
A customer is any person on any of the venues that wishes to dine at our
venue. Primary characteristics of a customer is that they have a type, a
time period of waiting for their food and a budget. Individual waiting
time and budget is usually determined by the type of the customer but in
some cases it might be different (See customer packs below for details).

A customer has a simple lifecycle. It first waits to be assigned to a
table by the TableManager. After that it goes to its table and orders a
food, influenced by the likes and dislikes of its type. If it waits too
long for the food to be made it prematurely leaves, wasting time and
resources for an unused food. Note that customer waits for the food to
be made, not delivered, if a dish is placed to the serving table of the
restaurant customer will not leave for waiting even if it takes two days
to be delivered. After getting its food the customer spends some time to
actually consume the dish and after eating it evaluates the food. Logic
for evaluation of the dishes is not implemented here as it is one of the
responsibilities of the game logic. After evaluation the customer waits
for other members of the pack to finish and then simply pay and leave.

CustomerPacks is a group of customers who wants to dine together at our
restaurant. Primary differences in behaviour between singular customers
and packs are they are seated together and the budget and the waiting
time of the pack members are updated to the maximum value in the pack.
Also if any of the pack member's dish gets made then entire pack will
wait their foods. These properties along with customers waiting other
pack memebers to pay emulate both the union of pack, by sychronized
waiting and leaving conditions, and the fact that groups of customers
are likely to spend more money per person.