Implement Advertisement class
Advertisement objects are where we store the advertisements given by the
player. They modify the popularity of certain demographics based on its
properties An advertisement has 4 primary properties. Quality represent
how much money is spent making the advert itself. Better quality means
the more effective and rememberable the advert is. Duration represent
how many days the advert will stay on the world. Platform means on which
platform the advert is. Better platforms reach to more people. Focus
represents the focused demographic of the advert. If you give an advert
focused on teens you would see most of the result in teens.