Implement Recipe, Order and Dish classes
Recipes are objects which holds a blueprint for a dish to be made. It
contains an ingredient list and a price both of which can be modified by
the player and a process list, describing which furniture a cook should
use for making the dish. Also each recipe has a unlock level and a bool
to hold if the player is serving this recipe at his/her venues.

Orders are objects that represents a dish requested by a customer. It
holds which recipe is requested by who, at what state the order is and
what is the current process element at the recipe process.

Dishes are recipes that are made by the cooks for some customers. Only
values that are hold is which recipe is used to make this dish and how
much of the dish is remaining (to keep track of the customer eating).