Implement Cook, Hawker and Waiter classes
Cooks have a simple routine to them. Wait for an order to arrive, follow
the process of the order and generate a dish for the recipe to the
serving table.Likewise Waiters have a simple routine to them. Wait for
a dish to be placed to the serving table, bring the dish to the customer
who ordered it.

Hawkers are a special type of employees that cannot be hired. They are
only present at the starting of the game when the player only has a food
stall. They are hybrid of Cooks and Waiters. They both prepare and
deliver the dish to the customer. However one significant difference is
that hawkers do not need to follow the process of the dish, they can
make anything at their stall without using any explicit cookers, ovens,
etc. It is implied that their stall has all the tools they need to make
a dish.